refactor: split layout/page_write.go — compression and encoding into page_write_encode.go
